Add adafactor optimizer.
This is a port of the Adafactor optimizer in flax.optim.
The core logic is implemented in `optax.scale_by_factored_rms`.

The flax.optim version of Adafactor supports a number of possible variants (e.g. from weight decay to momentum), `optax.adafactor` trivially supports the same set of variants, by just chaining the corresponding transforms for any of the options available in flax.optim.

This is my take on adafactor but I am happy for the Flax authors to take ownership of this component and restructure it however they see fit.

PiperOrigin-RevId: 374201573